Move InMemoryPodRecordRepository to test-utils
