+++ examples/keep-alive-framework/build.gradle
@@ -9,6 +9,8 @@ application {
 dependencies {
     implementation project(':core')
     implementation project(':persistence')
-    testCompile project(':test-utils')
+
+    // Use testCompile when we add implementation project(':persistence-zookeeper')
+    implementation project(':test-utils')
     implementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.1.7'
 }
