+++ persistence-zookeeper/src/test/scala/com/mesosphere/usi/repository/InMemoryRepositoryTest.scala
@@ -1,13 +1,14 @@
 package com.mesosphere.usi.repository
 
 import com.mesosphere.utils.UnitTest
+import com.mesosphere.utils.persistence.InMemoryPodRecordRepository
 import com.typesafe.scalalogging.StrictLogging
 
 class InMemoryRepositoryTest extends UnitTest with RepositoryBehavior with StrictLogging {
 
   "The in-memory pod record repository" should {
-    behave like podRecordStore(() => InMemoryPodRecordRepository.apply())
-    behave like podRecordReadAll(() => InMemoryPodRecordRepository.apply())
-    behave like podRecordDelete(() => InMemoryPodRecordRepository.apply())
+    behave like podRecordStore(() => InMemoryPodRecordRepository())
+    behave like podRecordReadAll(() => InMemoryPodRecordRepository())
+    behave like podRecordDelete(() => InMemoryPodRecordRepository())
   }
 }
